+++ uui/src/lib/shared/theme/palette/default-palettes.ts
@@ -48,7 +48,14 @@ export const UUI_PALETTES = {
   },
   dark: {
     background: '#0a0e2e',
-    onBackground: '#FFFFFF'
+    onBackground: '#FFFFFF',
+    textField: {
+      inactive: 'rgba(255, 255, 255, .6)',
+      hover: '#ffffff',
+      background: 'rgba(255, 255, 255, .1)',
+      backgroundHover: 'rgba(255, 255, 255, .3)',
+      bottomLineHover: 'rgba(255, 255, 255, 0.8)'
+    }
   },
   light: {
     background: '#FFFFFF',
@@ -56,6 +63,13 @@ export const UUI_PALETTES = {
     on: {
       background: '#000000',
       surface: '#000000'
+    },
+    textField: {
+      inactive: 'rgba(0, 0, 0, .6)',
+      hover: '#000000',
+      background: 'rgba(0, 0, 0, .1)',
+      backgroundHover: 'rgba(0, 0, 0, .3)',
+      bottomLineHover: 'rgba(0, 0, 0, 0.8)'
     }
   }
 };